requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
jinja2>=3.1.0
python-multipart>=0.0.6
groq>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
